refactor: Move recommendation view tools into folder.
