+++ src/tools/index.ts
@@ -22,7 +22,6 @@ import "./get-provider-resource";
 import "./get-recommendation-details";
 import "./get-recommendation-resource-details";
 import "./get-recommendation-resources";
-import "./get-recommendation-view";
 import "./get-report-notification";
 import "./get-team";
 import "./get-teams";
@@ -44,6 +43,7 @@ import "./list-tag-values";
 import "./list-tags";
 import "./list-unit-costs";
 import "./query-costs";
+import "./recommendation-views";
 import "./submit-user-feedback";
 import "./update-anomaly";
 import "./update-folder";
