+++ src/tools/recommendation-views/get-recommendation-view.test.ts
@@ -1,7 +1,7 @@
 import { type GetRecommendationViewResponse, pathEncode } from "@vantage-sh/vantage-client";
 import { expect } from "vitest";
 import tool from "./get-recommendation-view";
-import { requestsInOrder, testTool } from "./utils/testing";
+import { requestsInOrder, testTool } from "../utils/testing";
 
 const success: GetRecommendationViewResponse = {
   token: "rec_vw_be3f24eb1b5aabf6",
@@ -28,13 +28,6 @@ testTool(
         recommendation_view_token: "rec_vw_be3f24eb1b5aabf6",
       },
     },
-    {
-      name: "rejects empty recommendation_view_token",
-      data: {
-        recommendation_view_token: "",
-      },
-      expectedIssues: ["Too small: expected string to have >=1 characters"],
-    },
   ],
   [
     {
@@ -53,26 +46,6 @@ testTool(
       handler: async ({ callExpectingSuccess }) => {
         const res = await callExpectingSuccess({
           recommendation_view_token: "rec_vw_be3f24eb1b5aabf6",
-        });
-        expect(res).toEqual(success);
-      },
-    },
-    {
-      name: "encodes recommendation_view_token in endpoint",
-      apiCallHandler: requestsInOrder([
-        {
-          endpoint: `/v2/recommendation_views/${pathEncode("rec_vw_be3f24eb1b5aabf6/with space")}`,
-          params: {},
-          method: "GET",
-          result: {
-            ok: true,
-            data: success,
-          },
-        },
-      ]),
-      handler: async ({ callExpectingSuccess }) => {
-        const res = await callExpectingSuccess({
-          recommendation_view_token: "rec_vw_be3f24eb1b5aabf6/with space",
         });
         expect(res).toEqual(success);
       },
