+++ src/tools/recommendation-views/get-recommendation-view.ts
@@ -1,18 +1,14 @@
 import { pathEncode } from "@vantage-sh/vantage-client";
 import z from "zod/v4";
-import MCPUserError from "./structure/MCPUserError";
-import registerTool from "./structure/registerTool";
+import MCPUserError from "../structure/MCPUserError";
+import registerTool from "../structure/registerTool";
 
 const description = `
 Gets a specific recommendation view by its token.
-
-Use this tool when you already have a recommendation view token, such as one returned by Vantage or
-referenced in the Vantage console. The response includes the view's title, workspace, date range,
-provider and account scopes, regions, tags, creation timestamp, and creator.
 `.trim();
 
 const args = {
-  recommendation_view_token: z.string().min(1).describe("The recommendation view token to retrieve"),
+  recommendation_view_token: z.string().describe("The token of the recommendation view to retrieve"),
 };
 
 export default registerTool({
